requirements.txt
asgiref==3.8.1
Django==5.1.5
django-autoslug==1.9.9
django-ckeditor-5==0.2.15
django-cleanup==9.0.0
django-js-asset==3.0.1
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
sqlparse==0.5.3
whitenoise==6.8.2
